Move import state onto a new object
Leaving the ImportGLTF for the Blender interface only.
