+++ addons/io_scene_gltf_ksons/__init__.py
@@ -5,7 +5,6 @@ import struct
 import bpy
 from bpy.props import StringProperty, BoolProperty, FloatProperty, EnumProperty
 from bpy_extras.io_utils import ImportHelper
-from mathutils import Euler, Vector, Quaternion
 
 bl_info = {
     'name': "KSons' glTF 2.0 Importer",
@@ -33,8 +32,7 @@ EXTENSIONS = set((
     'MSFT_texture_dds',
 ))
 
-from . import animation, buffer, camera, material, mesh, scene, light, load, vnode, node
-
+from .importer import Importer
 
 class ImportGLTF(bpy.types.Operator, ImportHelper):
     """Load a glTF 2.0 file."""
@@ -144,27 +142,6 @@ class ImportGLTF(bpy.types.Operator, ImportHelper):
         default=True,
     )
 
-    def execute(self, context):
-        self.caches = {}
-
-        self.load_config()
-        self.set_conversions()
-
-        load.load(self)
-
-        # Precomputations
-        material.material_precomputation(self)
-        if self.import_animations:
-            animation.gather_animation_info(self)
-
-        vnode.create_vtree(self)
-        node.realize_vtree(self)
-        scene.create_blender_scenes(self)
-        if self.import_animations:
-            animation.add_animations(self)
-
-        return {'FINISHED'}
-
     def draw(self, context):
         layout = self.layout
         keywords = self.as_keywords()
@@ -200,72 +177,10 @@ class ImportGLTF(bpy.types.Operator, ImportHelper):
         col.prop(self, 'import_into_current_scene')
         col.prop(self, 'add_root')
 
-    def get(self, kind, id):
-        cache = self.caches.setdefault(kind, {})
-        if id in cache:
-            return cache[id]
-        else:
-            CREATE_FNS = {
-                'buffer': buffer.create_buffer,
-                'buffer_view': buffer.create_buffer_view,
-                'accessor': buffer.create_accessor,
-                'image': material.create_image,
-                'material': material.create_material,
-                'node_group': material.create_group,
-                'mesh': mesh.create_mesh,
-                'camera': camera.create_camera,
-                'light': light.create_light,
-            }
-            result = CREATE_FNS[kind](self, id)
-            if type(result) == dict and result.get('do_not_cache_me', False):
-                # Callee is requesting we not cache it
-                result = result['result']
-            else:
-                cache[id] = result
-            return result
-
-    def load_config(self):
-        """Load user-supplied options."""
-        keywords = self.as_keywords()
-        for opt in [
-            'import_into_current_scene', 'add_root',
-            'global_scale', 'axis_conversion',
-            'smooth_polys', 'split_meshes',
-            'import_animations', 'framerate', 'bone_rotation_mode',
-            'bone_rotation_axis', 'always_doublesided'
-        ]:
-            setattr(self, opt, keywords[opt])
-
-        if self.framerate <= 0:
-            self.framerate = bpy.context.scene.render.fps
-
-    def set_conversions(self):
-        global_scale = self.global_scale
-        axis_conversion = self.axis_conversion
-
-        if axis_conversion == 'BLENDER_UP':
-            def convert_translation(t):
-                return global_scale * Vector([t[0], -t[2], t[1]])
-
-            def convert_rotation(r):
-                return Quaternion([r[3], r[0], -r[2], r[1]])
-
-            def convert_scale(s):
-                return Vector([s[0], s[2], s[1]])
-
-        else:
-            def convert_translation(t):
-                return global_scale * Vector(t)
-
-            def convert_rotation(r):
-                return Quaternion([r[3], r[0], r[1], r[2]])
-
-            def convert_scale(s):
-                return Vector(s)
-
-        self.convert_translation = convert_translation
-        self.convert_rotation = convert_rotation
-        self.convert_scale = convert_scale
+    def execute(self, context):
+        imp = Importer(self.filepath, self.as_keywords())
+        imp.do_import()
+        return {'FINISHED'}
 
 
 # Add to a menu
